+++ R/elo.glm.R
@@ -1,59 +1,93 @@
 
 #' \code{elo.glm}
 #'
-#' Compute a logistic regression model for a matchup.
+#' Compute a (usually logistic) regression model for a matchup.
 #'
 #' @inheritParams elo.calc
-#' @param family,... Arguments passed to \code{\link[stats]{glm}}.
-#' @param rm.ties Logical, denoting whether to remove ties on the left-hand side.
-#' @return An object of class \code{c("elo.glm", "glm")}.
+#' @param family,weights,... Arguments passed to \code{\link[stats]{glm}}.
+#' @param running Logical, denoting whether to calculate "running" projected probabilities. If true, a model is fit for
+#'   group 1 on its own to predict group 2, then groups 1 and 2 to predict 3, then groups 1 through 3 to predict 4, etc.
+#'   Groups are determined in \code{formula}. Omitting a group term re-runs a glm model to predict each
+#'   observation (a potentially time-consuming operation!)
+#' @param skip Integer, denoting how many groups to skip before fitting the running models. This is helpful if
+#'   groups are small, where glm would have trouble converging for the first few groups. The predicted values are then
+#'   set to 0.5 for the skipped groups.
+#' @return An object of class \code{c("elo.glm", "glm")}. If \code{running==TRUE}, the class \code{"elo.glm.running"}
+#'   is prepended.
 #' @details
 #'   The formula syntax is the same as other \code{elo} functions. A data.frame
 #'   of indicator variables is built, where an entry is 1 if a team is home, 0 if
-#'   a team didn't play, and -1 if a team is a visitor. A \code{\link{glm}} model is then
-#'   run to predict wins.
+#'   a team didn't play, and -1 if a team is a visitor. Anything passed to \code{\link{adjust}()} in
+#'   \code{formula} is also put in the data.frame. A \code{\link{glm}} model is then
+#'   run to predict wins or margin of victory.
 #'
-#'   With this setup, usually one model term will be NA, as the input is linearly dependent.
-#'   Consider this team's skill to be zero relative to the other teams.
-#'   The intercept represents the home-field advantage.
+#'   With this setup, the intercept represents the home-field advantage. Neutral fields can be indicated
+#'   using the \code{\link{neutral}()} function, which sets the intercept to 0.
+#'
+#'   Note that any weights specified in \code{players()} will be ignored.
 #' @examples
 #' data(tournament)
-#' elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament)
-#'
-#' @seealso \code{\link{score}}, \code{\link{elo.model.frame}}
+#' elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament,
+#'   subset = points.Home != points.Visitor)
+#' elo.glm(mov(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament,
+#'   family = "gaussian")
+#' @seealso \code{\link[stats]{glm}}, \code{\link{summary.elo.glm}}, \code{\link{score}},
+#'   \code{\link{mov}}, \code{\link{elo.model.frame}}
 #' @name elo.glm
 NULL
 #> NULL
 
 #' @rdname elo.glm
 #' @export
-elo.glm <- function(formula, data, na.action, subset, family = "binomial", ..., rm.ties = TRUE)
+elo.glm <- function(formula, data, family = "binomial", weights, na.action, subset, ..., running = FALSE, skip = 0)
 {
   Call <- match.call()
   Call[[1L]] <- quote(elo::elo.model.frame)
-  Call$required.vars <- c("wins", "elos")
+  Call$required.vars <- c("wins", "elos", "group", "neutral", "weights")
   mf <- eval(Call, parent.frame())
+  if(nrow(mf) == 0) stop("No (non-missing) observations")
   Terms <- stats::terms(mf)
 
-  t1 <- mf$elo.A
-  t2 <- mf$elo.B
+  dat <- mf_to_wide(mf)
+  all.teams <- attr(dat, "all.teams")
 
-  if(is.numeric(t1) || is.numeric(t2)) stop("Neither team should be numeric")
-  if(!is.players(t1)) t1 <- players(t1)
-  if(anyNA(t1)) stop("NAs were found in team.A; check that it can be coerced to character.")
+  # find spanning set
+  QR <- qr(dat)
+  dat <- dat[QR$pivot[seq_len(QR$rank)]]
 
-  if(!is.players(t2)) t2 <- players(t2)
-  if(anyNA(t2)) stop("NAs were found in team.B; check that it can be coerced to character.")
-  all.teams <- sort(unique(c(as.character(t1), as.character(t2))))
-
-  dat <- lapply(all.teams, function(tm) (rowSums(t1 == tm) > 0) - (rowSums(t2 == tm) > 0))
-  names(dat) <- all.teams
   dat$wins.A <- mf$wins.A
-  dat <- structure(dat, class = "data.frame", row.names = c(NA_integer_, nrow(mf)))
+  grp <- mf$group
 
-  if(rm.ties) dat <- dat[dat$wins.A %in% 0:1, , drop = FALSE]
+  wts <- mf$weights
+  dat.glm <- stats::glm(wins.A ~ . - 1, data = dat, family = family, na.action = stats::na.pass, subset = NULL, weights = wts, ...)
+  dat.glm$teams <- all.teams
+  dat.glm$group <- grp
+  dat.glm$elo.terms <- Terms
+  dat.glm$na.action <- stats::na.action(mf)
+  dat.glm$outcome <- attr(mf, "outcome")
 
-  dat.glm <- stats::glm(wins.A ~ ., data = dat, family = family, ...)
+  if(running)
+  {
+    dat.mat <- as.matrix(dat[names(dat) != "wins.A"])
+    y <- dat$wins.A
 
-  structure(dat.glm, class = c("elo.glm", class(dat.glm)), rm.ties = rm.ties, all.teams = all.teams)
+    ftd <- rep(0, times = nrow(dat))
+    grp2 <- group_to_int(grp, skip)
+
+    for(i in setdiff(seq_len(max(grp2)), seq_len(skip)))
+    {
+      if(i == 1) next
+      sbst <- grp2 %in% 1:(i-1)
+
+      # tmpfit <- stats::glm(wins.A ~ . - 1, data = dat, subset = sbst, weights = wts, family = family)
+      # ftd[grp2 == i] <- predict(tmpfit, newdata = dat[grp2 == i, ], type = "link")
+
+      coeff <- stats::glm.fit(dat.mat[sbst, , drop = FALSE], y[sbst], wts[sbst], family = dat.glm$family,
+                              control = dat.glm$control)$coefficients
+      ftd[grp2 == i] <- apply(dat.mat[grp2 == i, , drop = FALSE], 1, function(x) sum(x * coeff, na.rm = TRUE))
+    }
+    dat.glm$running.values <- dat.glm$family$linkinv(ftd)
+  }
+
+  structure(dat.glm, class = c(if(running) "elo.running", "elo.glm", class(dat.glm)))
 }