+++ inst/doc/elo.R
@@ -45,6 +45,7 @@ elo.run(score(points.Home, points.Visitor) ~ team.Home + elo.Visitor +
 e <- elo.run(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
              data = tournament, k = 20)
 summary(e)
+rank.teams(e)
 
 ## ------------------------------------------------------------------------
 head(as.matrix(e))
@@ -77,4 +78,62 @@ elo.prob(~ elo.A + elo.B, data = dat)
 
 ## ------------------------------------------------------------------------
 elo.calc(wins.A ~ adjust(elo.A, 10) + elo.B + k(k), data = dat)
+
+## ------------------------------------------------------------------------
+e.winpct <- elo.winpct(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + group(week), data = tournament,
+                       subset = points.Home != points.Visitor) # to get rid of ties for now
+summary(e.winpct)
+rank.teams(e.winpct)
+predict(e.winpct, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE))
+
+tournament$neutral <- replace(rep(0, nrow(tournament)), 30:35, 1)
+summary(elo.winpct(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + neutral(neutral) + group(week),
+                   data = tournament, subset = points.Home != points.Visitor))
+
+## ------------------------------------------------------------------------
+e.winpct <- elo.winpct(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + group(week), data = tournament,
+                       subset = points.Home != points.Visitor, running = TRUE, skip = 5)
+summary(e.winpct)
+predict(e.winpct, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE)) # the same thing
+
+## ------------------------------------------------------------------------
+results <- elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + group(week), data = tournament,
+                   subset = points.Home != points.Visitor) # to get rid of ties for now
+summary(results)
+rank.teams(results)
+predict(results, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE))
+
+summary(elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + neutral(neutral) + group(week),
+                data = tournament, subset = points.Home != points.Visitor))
+
+## ------------------------------------------------------------------------
+results <- elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + group(week), data = tournament,
+                   subset = points.Home != points.Visitor, running = TRUE, skip = 5)
+summary(results)
+predict(results, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE)) # the same thing
+
+## ------------------------------------------------------------------------
+mc <- elo.markovchain(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament,
+                      subset = points.Home != points.Visitor, k = 0.7)
+summary(mc)
+rank.teams(mc)
+predict(mc, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE))
+summary(elo.markovchain(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + neutral(neutral),
+                        data = tournament, subset = points.Home != points.Visitor, k = 0.7))
+
+## ------------------------------------------------------------------------
+mc <- elo.markovchain(score(points.Home, points.Visitor) ~ team.Home + team.Visitor + group(week), data = tournament,
+                      subset = points.Home != points.Visitor, k = 0.7, running = TRUE, skip = 5)
+summary(mc)
+predict(mc, newdata = data.frame(team.Home = "Athletic Armadillos", team.Visitor = "Blundering Baboons", stringsAsFactors = FALSE)) # the same thing
+
+## ----eval=FALSE----------------------------------------------------------
+#  elo.markovchain(floor(wins.home) ~ team.home + team.visitor + k(ifelse(x > 0, rH(x), 1 - rH(x))))
+
+## ----eval=FALSE----------------------------------------------------------
+#  elo.markovchain(ifelse(home.points - visitor.points > h, 1, 0) ~ team.home + team.visitor + k(pmax(rH(x), 1 - rH(x))))
+
+## ------------------------------------------------------------------------
+summary(elo.glm(mov(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament,
+                family = "gaussian"))